backend/auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
python-multipart
fastapi-mail
python-dotenv 